Update dev dependencies to use global tools again
For a while we had to use outdated tools because we ran on python33,
and we used (or: had to use) type comments.  Since this is fixed, switch
back to glabl tools and always use the latest and greatest version.

Disable the newly added F824 from flake8 7.2 as `global` is also used
to make functions that access some globals visually stand out.